rework user preferences to create new forms
